# swagger.yaml
openapi: 3.0.4
info:
  title: API de Agenda de Ajustes Pós-Cirúrgicos
  description: |
    Sistema completo para gerenciamento de agenda de ajustes de produtos pós-cirúrgicos.

    ## Características
    - Autenticação JWT com controle de permissões
    - Cadastro de pessoas com diferentes papéis
    - Gestão de agendamentos com regras de negócio
    - Upload de fotos de ajustes
    - Sistema de logs para auditoria
    - Notificações via WhatsApp
    - Relatórios e dashboard

    ## Tipos de Usuário
    - **Administrador**: Acesso total ao sistema
    - **Secretária**: Cadastro de pacientes e agendamentos
    - **Costureira**: Registro de comparecimento e upload de fotos
    - **Médico**: Visualização de relatórios
    - **Paciente**: Recebe notificações (sem acesso ao sistema)

    ## Autenticação
    A maioria dos endpoints requer autenticação via JWT. 
    Obtenha um token através do endpoint `/api/auth/login` e inclua-o no header:
    ```
    Authorization: Bearer {seu_token}
    ```

  version: 1.0.0
  contact:
    name: Ramon David Souto
    email: rdssi@hotmail.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.sistema.com
    description: Servidor de produção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do login
  schemas:
...
    Error:
      type: object
      properties:
        sucesso:
          type: boolean
          example: false
        mensagem:
          type: string
          example: Erro na operação
        erro:
          type: string
          example: Detalhes do erro

    # Schema de Sucesso
    Success:
      type: object
      properties:
        sucesso:
          type: boolean
          example: true
        mensagem:
          type: string
          example: Operação realizada com sucesso

    # Schema de Login
...
    Foto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_produto_agendamento:
          type: integer
          example: 1
        des_caminho_arquivo:
          type: string
          example: uploads/ajustes/foto-1234567890.jpg
        nom_original:
          type: string
          example: foto_ajuste.jpg
        num_tamanho_bytes:
          type: integer
          example: 2048576
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    FotoList:
      type: object
      properties:
        sucesso:
          type: boolean
          example: true
        total:
          type: integer
          example: 5
        dados:
          type: array
          items:
            $ref: "#/components/schemas/Foto"

    # Schema de Log
...
  responses:
    UnauthorizedError:
      description: Token de autenticação ausente ou inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            sucesso: false
            mensagem: Não autorizado
            erro: Token não fornecido ou inválido
    ForbiddenError:
      description: Usuário não tem permissão para executar esta ação
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            sucesso: false
            mensagem: Acesso negado
            erro: Você não tem permissão para executar esta ação
paths:
...
  /api/produtos/agendamento:
    post:
      tags:
        - Produtos
      summary: Adicionar produto ao agendamento
      description: |
        Adiciona um produto à lista de produtos de um agendamento.

        **Permissões:**
        - Secretária: Pode adicionar
        - Costureira: Pode adicionar
        - Administrador: Pode adicionar
      operationId: adicionarProdutoAgendamento
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id_agendamento
                - id_produto_grade
              properties:
                id_agendamento:
                  type: integer
                  description: ID do agendamento
                  example: 1
                id_produto_grade:
                  type: integer
                  description: ID do produto na grade
                  example: 10
                des_ajuste_produto:
                  type: string
                  nullable: true
                  description: Descrição do ajuste necessário
                  example: Ajustar manga direita
                ind_status:
                  type: string
                  enum: [pendente, producao, concluido, cancelado]
                  default: pendente
                  example: pendente
      responses:
        "201":
          description: Produto adicionado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  mensagem:
                    type: string
                    example: Registro criado com sucesso
                  dados:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/produtos/agendamento/{id_agendamento}:
    get:
      tags:
        - Produtos
      summary: Listar produtos de um agendamento
      description: Lista todos os produtos vinculados a um agendamento, incluindo fotos
      operationId: listarProdutosAgendamento
      security:
        - bearerAuth: []
      parameters:
        - name: id_agendamento
          in: path
          required: true
          description: ID do agendamento
          schema:
            type: integer
      responses:
        "200":
          description: Lista de produtos do agendamento
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  total:
                    type: integer
                    example: 3
                  dados:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        id_produto_grade:
                          type: integer
                          example: 10
                        nom_produto:
                          type: string
                          example: Blusa Manga Longa
                        ref_produto:
                          type: string
                          example: BML-001
                        num_sku:
                          type: string
                          example: BML-001-P-AZ
                        cod_tamanho:
                          type: string
                          example: P
                        nom_cor_estampa:
                          type: string
                          example: Azul
                        des_ajuste_produto:
                          type: string
                          example: Ajustar manga direita
                        ind_status:
                          type: string
                          example: pendente
                        fotos:
                          type: array
                          items:
                            $ref: "#/components/schemas/Foto"
        "400":
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/produtos/{id}/status:
    patch:
      tags:
        - Produtos
      summary: Atualizar status de produto
      description: |
        Atualiza o status de um produto do agendamento.

        **Permissões:**
        - Costureira: Pode atualizar
        - Administrador: Pode atualizar
      operationId: atualizarStatusProduto
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do produto agendamento
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pendente, producao, concluido, cancelado]
                  description: Novo status do produto
                  example: producao
      responses:
        "200":
          description: Status atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
              example:
                sucesso: true
                mensagem: Registro atualizado com sucesso
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/produtos/{id_produto_agendamento}/fotos:
    post:
      tags:
        - Produtos
      summary: Upload de foto de ajuste (única)
      description: |
        Faz upload de uma foto do ajuste realizado.

        **Regras:**
        - Tamanho máximo: 5MB
        - Formatos aceitos: JPG, JPEG, PNG

        **Permissões:**
        - Costureira: Pode fazer upload
        - Administrador: Pode fazer upload
      operationId: uploadFotoAjuste
      security:
        - bearerAuth: []
      parameters:
        - name: id_produto_agendamento
          in: path
          required: true
          description: ID do produto agendamento
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - foto
              properties:
                foto:
                  type: string
                  format: binary
                  description: Arquivo de imagem (JPG, JPEG ou PNG)
      responses:
        "201":
          description: Foto enviada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  mensagem:
                    type: string
                    example: Foto enviada com sucesso
                  dados:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      nome_arquivo:
                        type: string
                        example: foto-1234567890.jpg
                      caminho:
                        type: string
                        example: uploads/ajustes/foto-1234567890.jpg
                      tamanho:
                        type: string
                        example: 2.5 MB
        "400":
          description: Erro no upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                arquivoGrande:
                  summary: Arquivo muito grande
                  value:
                    sucesso: false
                    mensagem: Arquivo muito grande
                    erro: Tamanho máximo permitido é 5 MB
                tipoInvalido:
                  summary: Tipo de arquivo inválido
                  value:
                    sucesso: false
                    mensagem: Arquivo inválido
                    erro: Tipo de arquivo não permitido. Tipos aceitos JPG, JPEG, PNG
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

    get:
      tags:
        - Produtos
      summary: Listar fotos de ajuste
      description: Lista todas as fotos de um produto do agendamento
      operationId: listarFotosAjuste
      security:
        - bearerAuth: []
      parameters:
        - name: id_produto_agendamento
          in: path
          required: true
          description: ID do produto agendamento
          schema:
            type: integer
      responses:
        "200":
          description: Lista de fotos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FotoList"
        "400":
          description: ID inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /api/produtos/{id_produto_agendamento}/fotos/multiplas:
    post:
      tags:
        - Produtos
      summary: Upload de múltiplas fotos de ajuste
      description: |
        Faz upload de múltiplas fotos do ajuste realizado.

        **Regras:**
        - Máximo: 10 fotos por vez
        - Tamanho máximo por foto: 5MB
        - Formatos aceitos: JPG, JPEG, PNG

        **Permissões:**
        - Costureira: Pode fazer upload
        - Administrador: Pode fazer upload
      operationId: uploadFotosMultiplas
      security:
        - bearerAuth: []
      parameters:
        - name: id_produto_agendamento
          in: path
          required: true
          description: ID do produto agendamento
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - fotos
              properties:
                fotos:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
                  description: Array de arquivos de imagem (máximo 10)
      responses:
        "201":
          description: Fotos enviadas com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  sucesso:
                    type: boolean
                    example: true
                  mensagem:
                    type: string
                    example: 3 foto(s) enviada(s) com sucesso
                  dados:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        nome_arquivo:
                          type: string
                          example: foto-1234567890.jpg
                        tamanho:
                          type: string
                          example: 2.5 MB
        "400":
          description: Erro no upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...